+++ src/synthesis/twytch_oscillators.cpp
@@ -21,8 +21,8 @@ namespace mopo {
   TwytchOscillators::TwytchOscillators() : Processor(kNumInputs, 2) { }
 
   void TwytchOscillators::process() {
-    Wave::Type waveform1 = static_cast<Wave::Type>(input(kOscillator1Waveform)->source->buffer[0]);
-    Wave::Type waveform2 = static_cast<Wave::Type>(input(kOscillator2Waveform)->source->buffer[0]);
+    Wave::Type waveform1 = static_cast<Wave::Type>(static_cast<int>(input(kOscillator1Waveform)->source->buffer[0]));
+    Wave::Type waveform2 = static_cast<Wave::Type>(static_cast<int>(input(kOscillator2Waveform)->source->buffer[0]));
 
     mopo_float base_phase1 = input(kOscillator1PhaseInc)->source->buffer[0];
     mopo_float base_phase2 = input(kOscillator2PhaseInc)->source->buffer[0];
